AIA/tts_engine/requirements.txt
piper-tts>=1.2.0
sounddevice>=0.4.6
pathlib